moved v3 specific code to v3 dir w/ burnettk
