+++ spiffworkflow-frontend/src/components/ProcessInstanceProgress.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { Loading } from '@carbon/react';
 
-import { CircularProgress } from '@mui/material';
 import {
   ErrorForDisplay,
   ProcessInstanceProgressResponse,
@@ -119,7 +119,14 @@ export default function ProcessInstanceProgress({
     if (smallSpinner) {
       style = { margin: '2x 5px 2px 2px' };
     }
-    return <CircularProgress style={style} />;
+    return (
+      <Loading
+        description="Active loading indicator"
+        withOverlay={false}
+        small={smallSpinner}
+        style={style}
+      />
+    );
   };
 
   const userMessage = (
